Moved option rendering to it’s own file
