+++ src/components/index.js
@@ -291,7 +291,7 @@ export {
   EuiSelectableSearch,
   EuiSelectableTemplateSitewide,
   EuiSelectableTemplateSitewideProps,
-  EuiSelectableTemplateSitewideSchema,
+  EuiSelectableTemplateSitewideOptionProps,
   EuiSelectableTemplateSitewideMetaData,
 } from './selectable';
