+++ src/components/selectable/selectable_templates/_index.scss
@@ -1,2 +1,2 @@
 @import 'variables';
-@import 'selectable_template_sitewide';
+@import 'selectable_template_sitewide_option';
